+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Changes prior to the next official version change will appear here.
 
 * Language Servers:
     * Add further file extensions considered by the language servers for Python (.pyi), JavaScript (.jsx) and TypeScript (.tsx, .jsx)
+    * Updated multilspy, adding support for Kotlin, Dart and C/C++ and several improvements.
     
 
 # 2025-04-07